[ui] basic catalog view for storage_kind tag
