+++ js_modules/dagster-ui/packages/ui-core/src/graph/OpTags.tsx
@@ -557,34 +557,6 @@ function coerceToStandardLabel(label: string) {
   return label.replace(/[ _-]/g, '').toLowerCase();
 }
 
-export const AssetComputeKindTag = ({
-  definition,
-  ...rest
-}: {
-  definition: {computeKind: string | null};
-  style: React.CSSProperties;
-  reduceColor?: boolean;
-  reduceText?: boolean;
-  reversed?: boolean;
-}) => {
-  if (!definition.computeKind) {
-    return null;
-  }
-  return (
-    <OpTags
-      {...rest}
-      tags={[
-        {
-          label: definition.computeKind,
-          onClick: () => {
-            window.requestAnimationFrame(() => document.dispatchEvent(new Event('show-kind-info')));
-          },
-        },
-      ]}
-    />
-  );
-};
-
 export const extractIconSrc = (knownTag: KnownTag | undefined) => {
   // Storybook imports SVGs are string but nextjs imports them as object.
   // This is a temporary work around until we can get storybook to import them the same way as nextjs
